Introduces custom hook to reduce code duplication.
