+++ web/libs/app-common/src/pages/AccountSettings/hooks/useHotkeys.ts
@@ -1,0 +1,329 @@
+import { useCallback, useEffect, useState } from "react";
+import { ToastType, useToast } from "@humansignal/ui";
+import { useAPI } from "apps/labelstudio/src/providers/ApiProvider";
+import { confirm } from "apps/labelstudio/src/components/Modal/Modal";
+import { DEFAULT_HOTKEYS } from "../sections/Hotkeys/defaults";
+
+// Types
+interface Hotkey {
+  id: string;
+  section: string;
+  element: string;
+  label: string;
+  key: string;
+  mac?: string;
+  active: boolean;
+}
+
+type HotkeySettings = {};
+
+interface ExportData {
+  hotkeys: Hotkey[];
+  settings: HotkeySettings;
+  exportedAt: string;
+  version: string;
+}
+
+interface ImportData {
+  hotkeys?: Hotkey[];
+  settings?: HotkeySettings;
+}
+
+interface SaveResult {
+  ok: boolean;
+  error?: string;
+  data?: any;
+}
+
+interface ApiResponse {
+  custom_hotkeys?: Record<string, { key: string; active: boolean; description?: string }>;
+  hotkey_settings?: HotkeySettings;
+  error?: string;
+}
+
+// Type the imported defaults and convert numeric ids to strings
+const typedDefaultHotkeys: Hotkey[] = DEFAULT_HOTKEYS.map((hotkey: any) => ({
+  ...hotkey,
+  id: String(hotkey.id), // Convert numeric id to string
+}));
+
+export const useHotkeys = () => {
+  const toast = useToast();
+  const [hotkeys, setHotkeys] = useState<Hotkey[]>([]);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const api = useAPI();
+
+  // Update hotkeys with custom settings
+  const updateHotkeysWithCustomSettings = useCallback(
+    (
+      defaultHotkeys: Hotkey[],
+      customHotkeys: Record<string, { key: string; active: boolean; description?: string }>,
+    ): Hotkey[] => {
+      return defaultHotkeys.map((hotkey: Hotkey) => {
+        // Create the lookup key format used in the API response (section:element)
+        const lookupKey = `${hotkey.section}:${hotkey.element}`;
+
+        // Check if there's a custom setting for this hotkey
+        if (customHotkeys[lookupKey]) {
+          const customSetting = customHotkeys[lookupKey];
+          // Create a new object with the default properties and override with custom ones
+          return {
+            ...hotkey,
+            key: customSetting.key,
+            active: customSetting.active,
+            // If description is provided in custom settings, use it as label
+            ...(customSetting.description && { label: customSetting.description }),
+          };
+        }
+
+        // If no custom setting exists, return the default hotkey unchanged
+        return hotkey;
+      });
+    },
+    [],
+  );
+
+  // Load hotkeys from API
+  const loadHotkeysFromAPI = useCallback(async () => {
+    try {
+      setIsLoading(true);
+
+      // Try to load from API first
+      const response = await api.callApi("hotkeys" as any);
+
+      if (response && (response as ApiResponse).custom_hotkeys) {
+        // Use API data
+        const apiResponse = response as ApiResponse;
+        const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, apiResponse.custom_hotkeys!);
+        setHotkeys(updatedHotkeys);
+      } else {
+        // Fallback to window.APP_SETTINGS
+        const customHotkeys = window.APP_SETTINGS?.user?.customHotkeys || {};
+        const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, customHotkeys);
+        setHotkeys(updatedHotkeys);
+      }
+    } catch (error) {
+      console.error("Error loading hotkeys from API:", error);
+
+      // Fallback to window.APP_SETTINGS on error
+      const customHotkeys = window.APP_SETTINGS?.user?.customHotkeys || {};
+      const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, customHotkeys);
+      setHotkeys(updatedHotkeys);
+
+      // Show non-blocking error notification
+      if (toast) {
+        toast.show({
+          message: "Could not load custom hotkeys from server, using cached settings",
+          type: ToastType.error,
+        });
+      }
+    } finally {
+      setIsLoading(false);
+    }
+  }, [api, toast, updateHotkeysWithCustomSettings]);
+
+  // Save hotkeys to API function (handles both save and reset operations)
+  const saveHotkeysToAPI = useCallback(
+    async (currentHotkeys: Hotkey[], currentSettings: HotkeySettings): Promise<SaveResult> => {
+      // Convert current hotkeys to API format
+      const customHotkeys: Record<string, { key: string; active: boolean }> = {};
+
+      // Process all current hotkeys (if empty, this results in reset)
+      currentHotkeys.forEach((hotkey: Hotkey) => {
+        const keyId = `${hotkey.section}:${hotkey.element}`;
+        customHotkeys[keyId] = {
+          key: hotkey.key,
+          active: hotkey.active,
+        };
+      });
+
+      const requestBody = {
+        custom_hotkeys: customHotkeys,
+        hotkey_settings: currentSettings,
+      };
+
+      try {
+        // Call the API to save/reset hotkeys and settings
+        const response = await api.callApi("updateHotkeys" as any, {
+          body: requestBody,
+        });
+
+        // Check for API-level errors
+        if (response?.error) {
+          return {
+            ok: false,
+            error: response.error,
+            data: response,
+          };
+        }
+
+        return {
+          ok: true,
+          error: undefined,
+          data: response,
+        };
+      } catch (error: unknown) {
+        const isReset = currentHotkeys.length === 0;
+        const operation = isReset ? "resetting" : "saving";
+        console.error(`Error ${operation} hotkeys:`, error);
+
+        // Provide more specific error messages
+        let errorMessage = `Failed to ${isReset ? "reset" : "save"} hotkeys`;
+        if (error && typeof error === "object" && "response" in error) {
+          const err = error as any;
+          // Server responded with error status
+          if (err.response?.status === 400) {
+            errorMessage = err.response.data?.error || `Invalid ${isReset ? "reset request" : "hotkeys configuration"}`;
+          } else if (err.response?.status === 401) {
+            errorMessage = "Authentication required";
+          } else if (err.response?.status >= 500) {
+            errorMessage = "Server error - please try again later";
+          }
+        } else if (error && typeof error === "object" && "request" in error) {
+          // Network error
+          errorMessage = "Network error - please check your connection";
+        }
+
+        return {
+          ok: false,
+          error: errorMessage,
+        };
+      }
+    },
+    [api],
+  );
+
+  // Handle resetting all hotkeys to defaults
+  const handleResetToDefaults = useCallback(() => {
+    confirm({
+      title: "Reset Hotkeys to Defaults?",
+      body: "Are you sure you want to reset all hotkeys and settings to their default values? This action cannot be undone.",
+      okText: "Reset to Defaults",
+      buttonLook: "negative",
+      style: { width: 500 },
+      onOk: async () => {
+        setIsLoading(true);
+
+        try {
+          // Reset hotkeys to defaults in the backend API (sets custom_hotkeys to {})
+          const result = await saveHotkeysToAPI([], {});
+
+          if (result.ok) {
+            if (toast) {
+              toast.show({
+                message: "All hotkeys and settings have been reset to defaults and saved",
+                type: ToastType.info,
+              });
+            }
+            // Update local state to reflect the reset
+            setHotkeys([...typedDefaultHotkeys]);
+          } else {
+            if (toast) {
+              toast.show({
+                message: `Failed to save reset hotkeys: ${result.error || "Unknown error"}`,
+                type: ToastType.error,
+              });
+            }
+          }
+        } catch (error: unknown) {
+          if (toast) {
+            const errorMessage = error instanceof Error ? error.message : "Unknown error";
+            toast.show({
+              message: `Error resetting hotkeys: ${errorMessage}`,
+              type: ToastType.error,
+            });
+          }
+        } finally {
+          setIsLoading(false);
+        }
+      },
+    });
+  }, [saveHotkeysToAPI, toast]);
+
+  // Handle exporting hotkeys
+  const handleExportHotkeys = useCallback(() => {
+    // Create export data including settings
+    const exportData: ExportData = {
+      hotkeys: hotkeys,
+      settings: {},
+      exportedAt: new Date().toISOString(),
+      version: "1.0",
+    };
+
+    // Create a JSON string of the export data
+    const exportJson = JSON.stringify(exportData, null, 2);
+
+    // Create a blob with the JSON
+    const blob = new Blob([exportJson], { type: "application/json" });
+    const url = URL.createObjectURL(blob);
+
+    // Create a temporary link and click it to download the file
+    const link = document.createElement("a");
+    link.href = url;
+    link.download = "hotkeys-export.json";
+    document.body.appendChild(link);
+    link.click();
+
+    // Clean up
+    document.body.removeChild(link);
+    URL.revokeObjectURL(url);
+
+    if (toast) {
+      toast.show({ message: "Hotkeys exported successfully", type: ToastType.info });
+    }
+  }, [hotkeys, toast]);
+
+  // Handle importing hotkeys
+  const handleImportHotkeys = useCallback(
+    async (importedData: ImportData | Hotkey[]) => {
+      try {
+        setIsLoading(true);
+
+        // Handle both old format (just hotkeys array) and new format (with settings)
+        const importedHotkeys = Array.isArray(importedData) ? importedData : importedData.hotkeys || [];
+
+        // Save all imported data to API
+        const result = await saveHotkeysToAPI(importedHotkeys, {});
+
+        if (!result.ok) {
+          throw new Error(result.error || "Failed to save imported hotkeys");
+        }
+
+        // Update local state
+        setHotkeys(importedHotkeys);
+
+        if (toast) {
+          toast.show({ message: "Hotkeys imported successfully", type: ToastType.info });
+        }
+
+        // Reload from API to ensure consistency
+        await loadHotkeysFromAPI();
+      } catch (error: unknown) {
+        if (toast) {
+          const errorMessage = error instanceof Error ? error.message : "Unknown error";
+          toast.show({ message: `Error importing hotkeys: ${errorMessage}`, type: ToastType.error });
+        }
+      } finally {
+        setIsLoading(false);
+      }
+    },
+    [saveHotkeysToAPI, loadHotkeysFromAPI, toast],
+  );
+
+  // Load hotkeys on hook mount
+  useEffect(() => {
+    loadHotkeysFromAPI();
+  }, [loadHotkeysFromAPI]);
+
+  return {
+    hotkeys,
+    setHotkeys,
+    isLoading,
+    setIsLoading,
+    loadHotkeysFromAPI,
+    saveHotkeysToAPI,
+    handleResetToDefaults,
+    handleExportHotkeys,
+    handleImportHotkeys,
+  };
+};
