+++ web/libs/app-common/src/pages/AccountSettings/sections/Hotkeys.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState, useCallback } from "react";
+import { useEffect, useState } from "react";
 import { IconWarning, ToastType, useToast } from "@humansignal/ui";
 
 // Shadcn UI components
@@ -13,7 +13,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@humansignal/shad/components/ui/dialog";
-import { useAPI } from "apps/labelstudio/src/providers/ApiProvider";
+// @ts-ignore
 import { confirm } from "apps/labelstudio/src/components/Modal/Modal";
 
 import { HotkeySection } from "./Hotkeys/Section";
@@ -21,6 +21,7 @@ import { ImportDialog } from "./Hotkeys/Import";
 import { KeyboardKey } from "./Hotkeys/Key";
 import { DEFAULT_HOTKEYS, HOTKEY_SECTIONS } from "./Hotkeys/defaults";
 import styles from "../AccountSettings.module.scss";
+import { useHotkeys } from "../hooks/useHotkeys";
 
 // Type definitions
 interface Hotkey {
@@ -80,271 +81,21 @@ interface ApiResponse {
 declare global {
   interface Window {
     DEFAULT_HOTKEYS: Hotkey[];
-    APP_SETTINGS?: {
-      user?: {
-        customHotkeys?: Record<string, { key: string; active: boolean }>;
-        hotkeySettings?: HotkeySettings;
-      };
-    };
   }
 }
 
 // Type the imported defaults
-const typedDefaultHotkeys = DEFAULT_HOTKEYS as Hotkey[];
+const typedDefaultHotkeys: Hotkey[] = DEFAULT_HOTKEYS.map((hotkey: any) => ({
+  ...hotkey,
+  id: String(hotkey.id), // Convert numeric id to string
+}));
 const typedHotkeySections = HOTKEY_SECTIONS as Section[];
 
 window.DEFAULT_HOTKEYS = typedDefaultHotkeys;
 
 export const HotkeysHeaderButtons = () => {
-  const toast = useToast();
-  const [hotkeys, setHotkeys] = useState<Hotkey[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
   const [importDialogOpen, setImportDialogOpen] = useState<boolean>(false);
-
-  const api = useAPI();
-
-  function updateHotkeysWithCustomSettings(
-    defaultHotkeys: Hotkey[],
-    customHotkeys: Record<string, { key: string; active: boolean; description?: string }>,
-  ): Hotkey[] {
-    return defaultHotkeys.map((hotkey: Hotkey) => {
-      // Create the lookup key format used in the API response (section:element)
-      const lookupKey = `${hotkey.section}:${hotkey.element}`;
-
-      // Check if there's a custom setting for this hotkey
-      if (customHotkeys[lookupKey]) {
-        const customSetting = customHotkeys[lookupKey];
-        // Create a new object with the default properties and override with custom ones
-        return {
-          ...hotkey,
-          key: customSetting.key,
-          active: customSetting.active,
-          // If description is provided in custom settings, use it as label
-          ...(customSetting.description && { label: customSetting.description }),
-        };
-      }
-
-      // If no custom setting exists, return the default hotkey unchanged
-      return hotkey;
-    });
-  }
-
-  // Load hotkeys from API
-  const loadHotkeysFromAPI = useCallback(async () => {
-    try {
-      setIsLoading(true);
-
-      // Try to load from API first
-      const response = await api.callApi("hotkeys" as any);
-
-      if (response && (response as ApiResponse).custom_hotkeys) {
-        // Use API data
-        const apiResponse = response as ApiResponse;
-        const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, apiResponse.custom_hotkeys!);
-        setHotkeys(updatedHotkeys);
-      } else {
-        // Fallback to window.APP_SETTINGS
-        const customHotkeys = window.APP_SETTINGS?.user?.customHotkeys || {};
-        const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, customHotkeys);
-        setHotkeys(updatedHotkeys);
-      }
-    } catch (error) {
-      console.error("Error loading hotkeys from API:", error);
-
-      // Fallback to window.APP_SETTINGS on error
-      const customHotkeys = window.APP_SETTINGS?.user?.customHotkeys || {};
-      const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, customHotkeys);
-      setHotkeys(updatedHotkeys);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [api]);
-
-  // Load hotkeys on component mount
-  useEffect(() => {
-    loadHotkeysFromAPI();
-  }, [loadHotkeysFromAPI]);
-
-  // Save hotkeys to API function (handles both save and reset operations)
-  const saveHotkeysToAPI = useCallback(
-    async (currentHotkeys: Hotkey[], currentSettings: HotkeySettings): Promise<SaveResult> => {
-      // Convert current hotkeys to API format
-      const customHotkeys: Record<string, { key: string; active: boolean }> = {};
-
-      // Process all current hotkeys (if empty, this results in reset)
-      currentHotkeys.forEach((hotkey: Hotkey) => {
-        const keyId = `${hotkey.section}:${hotkey.element}`;
-        customHotkeys[keyId] = {
-          key: hotkey.key,
-          active: hotkey.active,
-        };
-      });
-
-      const requestBody = {
-        custom_hotkeys: customHotkeys,
-        hotkey_settings: currentSettings,
-      };
-
-      try {
-        // Call the API to save/reset hotkeys and settings
-        const response = await api.callApi("updateHotkeys" as any, {
-          body: requestBody,
-        });
-
-        // Check for API-level errors
-        if (response?.error) {
-          return {
-            ok: false,
-            error: response.error,
-            data: response,
-          };
-        }
-
-        return {
-          ok: true,
-          error: undefined,
-          data: response,
-        };
-      } catch (error: unknown) {
-        const operation = "saving";
-        console.error(`Error ${operation} hotkeys:`, error);
-
-        // Provide more specific error messages
-        let errorMessage = "Failed to save hotkeys";
-        if (error && typeof error === "object" && "response" in error) {
-          const err = error as any;
-          // Server responded with error status
-          if (err.response?.status === 400) {
-            errorMessage = err.response.data?.error || "Invalid hotkeys configuration";
-          } else if (err.response?.status === 401) {
-            errorMessage = "Authentication required";
-          } else if (err.response?.status >= 500) {
-            errorMessage = "Server error - please try again later";
-          }
-        } else if (error && typeof error === "object" && "request" in error) {
-          // Network error
-          errorMessage = "Network error - please check your connection";
-        }
-
-        return {
-          ok: false,
-          error: errorMessage,
-        };
-      }
-    },
-    [api],
-  );
-
-  // Handle resetting all hotkeys to defaults
-  const handleResetToDefaults = () => {
-    confirm({
-      title: "Reset Hotkeys to Defaults?",
-      body: "Are you sure you want to reset all hotkeys and settings to their default values? This action cannot be undone.",
-      okText: "Reset to Defaults",
-      buttonLook: "negative",
-      style: { width: 500 },
-      onOk: async () => {
-        setIsLoading(true);
-
-        try {
-          // Reset hotkeys to defaults in the backend API (sets custom_hotkeys to {})
-          const result = await saveHotkeysToAPI([], {});
-
-          if (result.ok) {
-            if (toast) {
-              toast.show({
-                message: "All hotkeys and settings have been reset to defaults and saved",
-                type: ToastType.info,
-              });
-            }
-            // Update local state to reflect the reset
-            setHotkeys([...typedDefaultHotkeys]);
-          } else {
-            if (toast) {
-              toast.show({
-                message: `Failed to save reset hotkeys: ${result.error || "Unknown error"}`,
-                type: ToastType.error,
-              });
-            }
-          }
-        } catch (error: unknown) {
-          if (toast) {
-            const errorMessage = error instanceof Error ? error.message : "Unknown error";
-            toast.show({
-              message: `Error resetting hotkeys: ${errorMessage}`,
-              type: ToastType.error,
-            });
-          }
-        } finally {
-          setIsLoading(false);
-        }
-      },
-    });
-  };
-
-  // Handle exporting hotkeys
-  const handleExportHotkeys = () => {
-    // Create export data including settings
-    const exportData: ExportData = {
-      hotkeys: hotkeys,
-      settings: {},
-      exportedAt: new Date().toISOString(),
-      version: "1.0",
-    };
-
-    // Create a JSON string of the export data
-    const exportJson = JSON.stringify(exportData, null, 2);
-
-    // Create a blob with the JSON
-    const blob = new Blob([exportJson], { type: "application/json" });
-    const url = URL.createObjectURL(blob);
-
-    // Create a temporary link and click it to download the file
-    const link = document.createElement("a");
-    link.href = url;
-    link.download = "hotkeys-export.json";
-    document.body.appendChild(link);
-    link.click();
-
-    // Clean up
-    document.body.removeChild(link);
-    URL.revokeObjectURL(url);
-
-    if (toast) {
-      toast.show({ message: "Hotkeys exported successfully", type: ToastType.info });
-    }
-  };
-
-  // Handle importing hotkeys
-  const handleImportHotkeys = async (importedData: ImportData | Hotkey[]) => {
-    try {
-      setIsLoading(true);
-
-      // Handle both old format (just hotkeys array) and new format (with settings)
-      const importedHotkeys = Array.isArray(importedData) ? importedData : importedData.hotkeys || [];
-
-      // Save all imported data to API
-      const result = await saveHotkeysToAPI(importedHotkeys, {});
-
-      if (!result.ok) {
-        throw new Error(result.error || "Failed to save imported hotkeys");
-      }
-
-      if (toast) {
-        toast.show({ message: "Hotkeys imported successfully", type: ToastType.info });
-      }
-
-      // Reload the page to reflect changes
-      window.location.reload();
-    } catch (error: unknown) {
-      if (toast) {
-        const errorMessage = error instanceof Error ? error.message : "Unknown error";
-        toast.show({ message: `Error importing hotkeys: ${errorMessage}`, type: ToastType.error });
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const { handleResetToDefaults, handleExportHotkeys, handleImportHotkeys } = useHotkeys();
 
   return (
     <>
@@ -368,8 +119,6 @@ export const HotkeysHeaderButtons = () => {
 
 export const HotkeysManager = () => {
   const toast = useToast();
-  const [hotkeys, setHotkeys] = useState<Hotkey[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
   const [editingHotkeyId, setEditingHotkeyId] = useState<string | null>(null);
   const [globalEnabled, setGlobalEnabled] = useState<boolean>(true);
   const [dirtyState, setDirtyState] = useState<DirtyState>({});
@@ -381,153 +130,19 @@ export const HotkeysManager = () => {
     conflictingHotkeys: [],
   });
 
-  const api = useAPI();
+  // Use the shared hook for common functionality
+  const { hotkeys, setHotkeys, isLoading, setIsLoading, saveHotkeysToAPI } = useHotkeys();
 
   // Check if a hotkey conflicts with others globally
   const getGlobalDuplicates = (hotkeyId: string, newKey: string): Hotkey[] => {
     return hotkeys.filter((h: Hotkey) => h.id !== hotkeyId && h.key === newKey);
   };
 
-  // Save hotkeys to API function (handles both save and reset operations)
-  const saveHotkeysToAPI = useCallback(
-    async (currentHotkeys: Hotkey[], currentSettings: HotkeySettings): Promise<SaveResult> => {
-      // Convert current hotkeys to API format
-      const customHotkeys: Record<string, { key: string; active: boolean }> = {};
-
-      // Process all current hotkeys (if empty, this results in reset)
-      currentHotkeys.forEach((hotkey: Hotkey) => {
-        const keyId = `${hotkey.section}:${hotkey.element}`;
-        customHotkeys[keyId] = {
-          key: hotkey.key,
-          active: hotkey.active,
-        };
-      });
-
-      const requestBody = {
-        custom_hotkeys: customHotkeys,
-        hotkey_settings: currentSettings,
-      };
-
-      try {
-        // Call the API to save/reset hotkeys and settings
-        const response = await api.callApi("updateHotkeys" as any, {
-          body: requestBody,
-        });
-
-        // Check for API-level errors
-        if (response?.error) {
-          return {
-            ok: false,
-            error: response.error,
-            data: response,
-          };
-        }
-
-        return {
-          ok: true,
-          error: undefined,
-          data: response,
-        };
-      } catch (error: unknown) {
-        const isReset = currentHotkeys.length === 0;
-        const operation = isReset ? "resetting" : "saving";
-        console.error(`Error ${operation} hotkeys:`, error);
-
-        // Provide more specific error messages
-        let errorMessage = `Failed to ${isReset ? "reset" : "save"} hotkeys`;
-        if (error && typeof error === "object" && "response" in error) {
-          const err = error as any;
-          // Server responded with error status
-          if (err.response?.status === 400) {
-            errorMessage = err.response.data?.error || `Invalid ${isReset ? "reset request" : "hotkeys configuration"}`;
-          } else if (err.response?.status === 401) {
-            errorMessage = "Authentication required";
-          } else if (err.response?.status >= 500) {
-            errorMessage = "Server error - please try again later";
-          }
-        } else if (error && typeof error === "object" && "request" in error) {
-          // Network error
-          errorMessage = "Network error - please check your connection";
-        }
-
-        return {
-          ok: false,
-          error: errorMessage,
-        };
-      }
-    },
-    [api],
-  );
-
-  function updateHotkeysWithCustomSettings(
-    defaultHotkeys: Hotkey[],
-    customHotkeys: Record<string, { key: string; active: boolean; description?: string }>,
-  ): Hotkey[] {
-    return defaultHotkeys.map((hotkey: Hotkey) => {
-      // Create the lookup key format used in the API response (section:element)
-      const lookupKey = `${hotkey.section}:${hotkey.element}`;
-
-      // Check if there's a custom setting for this hotkey
-      if (customHotkeys[lookupKey]) {
-        const customSetting = customHotkeys[lookupKey];
-        // Create a new object with the default properties and override with custom ones
-        return {
-          ...hotkey,
-          key: customSetting.key,
-          active: customSetting.active,
-          // If description is provided in custom settings, use it as label
-          ...(customSetting.description && { label: customSetting.description }),
-        };
-      }
-
-      // If no custom setting exists, return the default hotkey unchanged
-      return hotkey;
-    });
-  }
-
-  // Load hotkeys from API
-  const loadHotkeysFromAPI = useCallback(async () => {
-    try {
-      setIsLoading(true);
-
-      // Try to load from API first
-      const response = await api.callApi("hotkeys" as any);
-
-      if (response && (response as ApiResponse).custom_hotkeys) {
-        // Use API data
-        const apiResponse = response as ApiResponse;
-        const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, apiResponse.custom_hotkeys!);
-        setHotkeys(updatedHotkeys);
-      } else {
-        // Fallback to window.APP_SETTINGS
-        const customHotkeys = window.APP_SETTINGS?.user?.customHotkeys || {};
-        const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, customHotkeys);
-        setHotkeys(updatedHotkeys);
-      }
-    } catch (error) {
-      console.error("Error loading hotkeys from API:", error);
-
-      // Fallback to window.APP_SETTINGS on error
-      const customHotkeys = window.APP_SETTINGS?.user?.customHotkeys || {};
-      const updatedHotkeys = updateHotkeysWithCustomSettings(typedDefaultHotkeys, customHotkeys);
-      setHotkeys(updatedHotkeys);
-
-      // Show non-blocking error notification
-      if (toast) {
-        toast.show({
-          message: "Could not load custom hotkeys from server, using cached settings",
-          type: ToastType.error,
-        });
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  }, [api, toast]);
-
-  // Load hotkeys on component mount
+  // Update globalEnabled state when hotkeys change
   useEffect(() => {
-    loadHotkeysFromAPI();
-  }, [loadHotkeysFromAPI]);
+    const allEnabled = hotkeys.every((hotkey: Hotkey) => hotkey.active);
+    setGlobalEnabled(allEnabled);
+  }, [hotkeys]);
 
   // Handle toggling a single hotkey
   const handleToggleHotkey = (hotkeyId: string) => {
@@ -555,7 +170,7 @@ export const HotkeysManager = () => {
     setGlobalEnabled(allEnabled);
   };
 
-  // Handle resetting all hotkeys to defaults
+  // Handle resetting all hotkeys to defaults (enhanced version with dirty state management)
   const handleResetToDefaults = async () => {
     const hasChanges = hasUnsavedChanges || dirtyState.settings;
     const confirmMessage = hasChanges
@@ -741,40 +356,7 @@ export const HotkeysManager = () => {
     }
   };
 
-  // Handle exporting hotkeys
-  const handleExportHotkeys = () => {
-    // Create export data including settings
-    const exportData: ExportData = {
-      hotkeys: hotkeys,
-      settings: {},
-      exportedAt: new Date().toISOString(),
-      version: "1.0",
-    };
-
-    // Create a JSON string of the export data
-    const exportJson = JSON.stringify(exportData, null, 2);
-
-    // Create a blob with the JSON
-    const blob = new Blob([exportJson], { type: "application/json" });
-    const url = URL.createObjectURL(blob);
-
-    // Create a temporary link and click it to download the file
-    const link = document.createElement("a");
-    link.href = url;
-    link.download = "hotkeys-export.json";
-    document.body.appendChild(link);
-    link.click();
-
-    // Clean up
-    document.body.removeChild(link);
-    URL.revokeObjectURL(url);
-
-    if (toast) {
-      toast.show({ message: "Hotkeys exported successfully", type: ToastType.info });
-    }
-  };
-
-  // Handle importing hotkeys
+  // Enhanced import handler that manages dirty state
   const handleImportHotkeys = async (importedData: ImportData | Hotkey[]) => {
     try {
       setIsLoading(true);
@@ -795,9 +377,6 @@ export const HotkeysManager = () => {
       if (!result.ok) {
         throw new Error(result.error || "Failed to save imported hotkeys");
       }
-
-      // Reload from API to ensure consistency
-      await loadHotkeysFromAPI();
 
       // Reset dirty state
       setDirtyState({});
